// src/types.ts
export interface OTInventoryItem {
  id: string;
  code?: string;
  name: string;
  category: 'Surgical Kit' | 'Disposable' | 'Anesthesia Drug' | 'Suture' | 'Implant';
  stock: number;
  unit: string;
  minStockLevel: number;
  mrp: number;
  purchasePrice: number;
  batchNumber?: string;
  expiryDate?: string;
}